Updated to support Maven
