+++ .gitignore
@@ -3,3 +3,6 @@
 .classpath
 .settings/
 .project
+/*.iml
+/.idea/
+/lib/*
