--- src/plugin.yml
@@ -1,12 +1,0 @@
-name: Commandspy
-main: me.korikisulda.commandspy.commandspy
-version: 2.9.3
-author: korikisulda
-description: >
-             A plugin to tell you what everyone's doing. mostly.
-commands:
-    commandspy:
-      description: Access commandspy subcommands
-      usage: /commandspy help
-      aliases: [c, cs]
-softdepend: [WorldEdit]
